class2/pyproject.toml
[project]
name = "class2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
]
